impl

+++ src/main/java/io/github/mateusz00/configuration/SecurityConstants.java
@@ -17,5 +17,5 @@ public class SecurityConstants {
     @Value("${jwt.prefix}")
     private String tokenPrefix;
     @Value("${jwt.expiration}")
-    private Integer expirationTime;
+    private Long expirationTime;
 }
